# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Api",
        "description": "Demo Library Api<br><br>Develped by <strong>Rafael Lessa de Castro Xavier</strong>",
        "contact": {
            "email": "rafaellessacastro@hotmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/rent/store": {
            "post": {
                "tags": [
                    "Rent"
                ],
                "summary": "Create Rent",
                "operationId": "91d64b773ad9343931d6a12072680608",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "devolution_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2022-03-28"
                                            },
                                            "client_id": {
                                                "type": "number",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "devolution_date": "2022-03-28",
                                    "client_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "devolution_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-03-28"
                                        },
                                        "client_id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "client_full_name": {
                                            "type": "string",
                                            "example": "Full name"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "opened"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rent/find/{id}": {
            "get": {
                "tags": [
                    "Rent"
                ],
                "summary": "Get Rent Details",
                "operationId": "b78da48db8ddd19201813d4274eee582",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "devolution_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-03-28"
                                        },
                                        "client_id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "client_full_name": {
                                            "type": "string",
                                            "example": "Full name"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "opened"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rent not found."
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rent/update/{id}": {
            "put": {
                "tags": [
                    "Rent"
                ],
                "summary": "Update Rent",
                "operationId": "b244102f8389f55350999a48aeabcc83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "devolution_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-03-28"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/rent/delete/{id}": {
            "delete": {
                "tags": [
                    "Rent"
                ],
                "summary": "Delete Rent",
                "operationId": "1e75efe8bad00dc6ac796d46a324b95b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "Not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rent not found."
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rent/close/{id}": {
            "delete": {
                "tags": [
                    "Rent"
                ],
                "summary": "Close Rent",
                "operationId": "ab7fbd87947c37a9624c60e856a9989c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The rent is closed successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rent not found."
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}